boards/arm/stm32f7: move ROMFS logic to common directory
move ROMFS logic to common directory
