+++ LICENSE
@@ -3355,12 +3355,8 @@ boards/arm/stm32/omnibusf4/src/stm32_romfs.h
 boards/arm/stm32/omnibusf4/src/stm32_romfs_initialize.c
 boards/arm/stm32/stm32f4discovery/src/stm32_romfs.h
 boards/arm/stm32/stm32f4discovery/src/stm32_romfs_initialize.c
-boards/arm/stm32f7/nucleo-f722ze/src/stm32_romfs.h
-boards/arm/stm32f7/nucleo-f722ze/src/stm32_romfs_initialize.c
-boards/arm/stm32f7/nucleo-f746zg/src/stm32_romfs.h
-boards/arm/stm32f7/nucleo-f746zg/src/stm32_romfs_initialize.c
-boards/arm/stm32f7/nucleo-f767zi/src/stm32_romfs.h
-boards/arm/stm32f7/nucleo-f767zi/src/stm32_romfs_initialize.c
+boards/arm/stm32f7/common/include/stm32_romfs.h
+boards/arm/stm32f7/common/src/stm32_romfs_initialize.c
 =============================================================
   Copyright (C) 2017 Tomasz Wozniak. All rights reserved.
