+++ boards/Kconfig
@@ -4604,6 +4604,9 @@ endif
 if ARCH_CHIP_STM32
 source "boards/arm/stm32/common/Kconfig"
 endif
+if ARCH_CHIP_STM32F7
+source "boards/arm/stm32f7/common/Kconfig"
+endif
 if ARCH_CHIP_STM32F0L0G0
 source "boards/arm/stm32f0l0g0/common/Kconfig"
 endif
