+++ boards/arm/stm32f7/common/Kconfig
@@ -3,3 +3,25 @@
 # see the file kconfig-language.txt in the NuttX tools repository.
 #
 
+config STM32F7_ROMFS
+	bool "Automount baked-in ROMFS image"
+	default n
+	depends on FS_ROMFS
+	---help---
+		Select STM32F7_ROMFS_IMAGEFILE, STM32F7_ROMFS_DEV_MINOR, STM32F7_ROMFS_MOUNTPOINT
+
+config STM32F7_ROMFS_DEV_MINOR
+	int "Minor for the block device backing the data"
+	depends on STM32F7_ROMFS
+	default 64
+
+config STM32F7_ROMFS_MOUNTPOINT
+	string "Mountpoint of the custom romfs image"
+	depends on STM32F7_ROMFS
+	default "/rom"
+
+config STM32F7_ROMFS_IMAGEFILE
+	string "ROMFS image file to include into build"
+	depends on STM32F7_ROMFS
+	default "../../../rom.img"
+
