+++ boards/arm/stm32f7/common/src/Make.defs
@@ -41,6 +41,10 @@ CSRCS += stm32_cansock_setup.c
 endif
 endif
 
+ifeq ($(CONFIG_STM32F7_ROMFS),y)
+CSRCS += stm32_romfs_initialize.c
+endif
+
 DEPPATH += --dep-path src
 VPATH += :src
 CFLAGS += ${INCDIR_PREFIX}$(TOPDIR)$(DELIM)arch$(DELIM)$(CONFIG_ARCH)$(DELIM)src$(DELIM)board$(DELIM)src
