+++ boards/arm/stm32f7/nucleo-f746zg/Kconfig
@@ -5,28 +5,6 @@
 
 if ARCH_BOARD_NUCLEO_F746ZG
 
-config STM32_ROMFS
-	bool "Automount baked-in ROMFS image"
-	default n
-	depends on FS_ROMFS
-	---help---
-		Select STM32_ROMFS_IMAGEFILE, STM32_ROMFS_DEV_MINOR, STM32_ROMFS_MOUNTPOINT
-
-config STM32_ROMFS_DEV_MINOR
-	int "Minor for the block device backing the data"
-	depends on STM32_ROMFS
-	default 64
-
-config STM32_ROMFS_MOUNTPOINT
-	string "Mountpoint of the custom romfs image"
-	depends on STM32_ROMFS
-	default "/rom"
-
-config STM32_ROMFS_IMAGEFILE
-	string "ROMFS image file to include into build"
-	depends on STM32_ROMFS
-	default "../../../rom.img"
-
 choice
 	prompt "Select Console wiring."
 	default NUCLEO_F746ZG_CONSOLE_ARDUINO
